In this push we have the second waveform working perfectly. Additionally, the cluster view shows the spikes from the second waveform with a different symbol and color.

+++ src/waveform2view.h
@@ -57,7 +57,7 @@ public:
     friend class Waveform2Thread;
 
     Waveform2View(KlustersDoc& doc, KlustersView& view, const QColor &backgroundColor, int acquisitionGain, const QList<int> &positions, QStatusBar * statusBar, QWidget* parent=0,
-                 bool isTimeFrameMode = false, long start = 0, long timeFrameWidth = 0, long nbSpkToDisplay =0, bool overLay = false, bool mean = false, int minSpikeDiff =1,
+                 bool isTimeFrameMode = false, long start = 0, long timeFrameWidth = 0, long nbSpkToDisplay =0, bool overLay = false, bool mean = false, double minSpikeDiff =1,
                  const char* name=0, int minSize = 50, int maxSize = 4000, int windowTopLeft = -500,
                  int windowBottomRight = 1001, int border = 0);
     ~Waveform2View();
@@ -214,7 +214,7 @@ public Q_SLOTS:
   */
     void setDisplayNbSpikes(long nbSpikes);
 
-    void setMinSpikeDiff(long MinSpkDiff);
+    void setMinSpikeDiff(double MinSpkDiff);
 
     /**Enables the caller to know if there is any thread running launch by the view.*/
     bool isThreadsRunning() const;
@@ -391,7 +391,7 @@ private:
     /**The number of spikes to display in sample mode.*/
     long nbSpkToDisplay;
 
-    int minSpikeDiff;
+    double minSpikeDiff;
 
     /**True if the view has been zoomed, false ohterwise.*/
     bool isZoomed;
